requirements.txt
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
APScheduler==3.10.4
asttokens==2.4.1
cachetools==5.3.2
certifi==2023.11.17
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
deprecation==2.1.0
distlib==0.3.8
distro==1.9.0
executing==2.0.1
fastapi==0.108.0
filelock==3.13.1
google-auth==2.26.2
google-auth-oauthlib==1.2.0
gotrue==2.1.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
identify==2.5.33
idna==3.6
iniconfig==2.0.0
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.0
matplotlib-inline==0.1.6
mypy==1.8.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nodeenv==1.8.0
numpy==1.26.3
oauthlib==3.2.2
openai==1.7.0
packaging==23.2
pandas==2.1.4
pandas-stubs==2.1.4.231227
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.3.0
postgrest==0.13.1
pre-commit==3.6.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyTelegramBotAPI==4.14.1
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.6.1
realtime==1.0.2
reportlab==4.0.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sentry-sdk==1.39.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.32.0.post1
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.1
supafunc==0.3.3
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
types-openpyxl==3.1.0.20240106
types-pytz==2023.3.1.1
types-requests==2.31.0.20240106
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.25.0
virtualenv==20.25.0
wcwidth==0.2.12
websockets==11.0.3
